[script] Fix bug in tap2wake

+++ stage2/03-crankshaft-base/files/opt/crankshaft/service_tap2wake.sh
@@ -9,7 +9,7 @@ readlines() {
         check=$(echo $line | grep "^Event:")
         if [ -n "$check" ]; then
             systemctl stop disconnect.service
-        if [ ! -f /tmp/dev_mode_enabled ]; then
+        if [ ! -f /tmp/dev_mode_enabled ] && [ ! -f /tmp/android_device ]; then
             systemctl stop disconnect.timer
             systemctl start disconnect.timer
         fi
